+++ fpsim/sim.py
@@ -8,7 +8,7 @@ import pylab as pl
 import seaborn as sns
 import sciris as sc
 import pandas as pd
-from . import defaults as fpd
+from . import parameters as fpp
 from . import utils as fpu
 from . import base as fpb
 
@@ -40,7 +40,7 @@ class People(fpb.BasePeople):
         # Initialization
         super().__init__()
         self.pars = pars # Set parameters
-        d = sc.mergedicts(fpd.person_defaults, kwargs) # d = defaults
+        d = sc.mergedicts(fpp.person_defaults, kwargs) # d = defaults
         if n is None:
             n = int(self.pars['n_agents'])
 
@@ -108,11 +108,11 @@ class People(fpb.BasePeople):
         m = len(method_map)
         switching_events = np.zeros((m, m), dtype=int)
         switching_events_ages = {}
-        for key in fpd.method_age_map.keys():
+        for key in fpp.method_age_map.keys():
             switching_events_ages[key] = np.zeros((m, m), dtype=int)
 
         # Method switching depends both on agent age and also on their current method, so we need to loop over both
-        for key,(age_low, age_high) in fpd.method_age_map.items():
+        for key,(age_low, age_high) in fpp.method_age_map.items():
             match_low  = (self.age >= age_low) # CK: TODO: refactor into single method
             match_high = (self.age <  age_high)
             match_low_high = match_low * match_high
@@ -136,7 +136,7 @@ class People(fpb.BasePeople):
 
         if self.pars['track_switching']:
             self.step_results_switching['annual'] += switching_events # CK: TODO: remove this extra result and combine with step_results
-            for key in fpd.method_age_map.keys():
+            for key in fpp.method_age_map.keys():
                 self.step_results['switching_annual'][key] += switching_events_ages[key]
 
         return
@@ -162,7 +162,7 @@ class People(fpb.BasePeople):
         m = len(methods_map)
         switching_events = np.zeros((m, m), dtype=int)
         switching_events_ages = {}
-        for key in fpd.method_age_map.keys():
+        for key in fpp.method_age_map.keys():
             switching_events_ages[key] = np.zeros((m, m), dtype=int)
 
         postpartum1 = (self.postpartum_dur == 0)
@@ -170,7 +170,7 @@ class People(fpb.BasePeople):
 
         # In first time step after delivery, choice is by age but not previous method (since just gave birth)
         # All women are coming from birth and on no method to start, either will stay on no method or initiate a method
-        for key, (age_low, age_high) in fpd.method_age_map.items():
+        for key, (age_low, age_high) in fpp.method_age_map.items():
             match_low = (self.age >= age_low)
             match_high = (self.age < age_high)
             low_parity = (self.parity < self.pars['high_parity'])
@@ -205,7 +205,7 @@ class People(fpb.BasePeople):
         # At 6 months, choice is by previous method and by age
         # Allow initiation, switching, or discontinuing with matrix at 6 months postpartum
         # Transitional probabilities are for 5 months, 1-6 months after delivery from DHS data
-        for key,(age_low, age_high) in fpd.method_age_map.items():
+        for key,(age_low, age_high) in fpp.method_age_map.items():
             match_low  = (self.age >= age_low)
             match_high = (self.age <  age_high)
             match_postpartum_age = self.postpartum * postpartum6 * match_low * match_high
@@ -227,7 +227,7 @@ class People(fpb.BasePeople):
 
         if self.pars['track_switching']:
             self.step_results_switching['postpartum'] += switching_events
-            for key in fpd.method_age_map.keys():
+            for key in fpp.method_age_map.keys():
                 self.step_results['switching_postpartum'][key] += switching_events_ages[key]
 
         return
@@ -244,7 +244,7 @@ class People(fpb.BasePeople):
             pp.update_method_pp() # Update method for
 
             age_diff = non_pp.ceil_age - non_pp.age
-            whole_years = ((age_diff < (1/fpd.mpy)) * (age_diff > 0))
+            whole_years = ((age_diff < (1/fpp.mpy)) * (age_diff > 0))
             birthdays = non_pp.filter(whole_years)
             birthdays.update_method()
 
@@ -351,7 +351,7 @@ class People(fpb.BasePeople):
         # Adjust for probability of exposure to pregnancy episode at this timestep based on age and parity - encapsulates background factors - experimental and tunable
         preg_probs *= self.pars['exposure_factor']
         preg_probs *= self.pars['exposure_age'][all_ppl.int_age_clip]
-        preg_probs *= self.pars['exposure_parity'][np.minimum(all_ppl.parity, fpd.max_parity)]
+        preg_probs *= self.pars['exposure_parity'][np.minimum(all_ppl.parity, fpp.max_parity)]
 
         # Use a single binomial trial to check for conception successes this month
         conceived = active.binomial(preg_probs[active.inds], as_filter=True)
@@ -436,7 +436,7 @@ class People(fpb.BasePeople):
 
         # Count the state of the agent for postpartum -- # TOOD: refactor, what is this loop doing?
         pp = self.filter(self.postpartum)
-        for key,(pp_low, pp_high) in fpd.postpartum_map.items():
+        for key,(pp_low, pp_high) in fpp.postpartum_map.items():
             this_pp_bin = pp.filter((pp.postpartum_dur >= pp_low) * (pp.postpartum_dur <  pp_high))
             self.step_results[key] += len(this_pp_bin)
         pp.postpartum_dur += self.pars['timestep']
@@ -543,7 +543,7 @@ class People(fpb.BasePeople):
         self.step_results['total_births'] = len(stillborn) + self.step_results['births']
 
         live_age = live.age
-        for key, (age_low, age_high) in fpd.age_bin_map.items():
+        for key, (age_low, age_high) in fpp.age_bin_map.items():
             birth_bins = np.sum((live_age >= age_low) * (live_age < age_high))
             self.step_results['birth_bins'][key] += birth_bins
 
@@ -575,7 +575,7 @@ class People(fpb.BasePeople):
 
     def update_age(self):
         '''Advance age in the simulation'''
-        self.age += self.pars['timestep'] / fpd.mpy  # Age the person for the next timestep
+        self.age += self.pars['timestep'] / fpp.mpy  # Age the person for the next timestep
         self.age = np.minimum(self.age, self.pars['max_age'])
         return
 
@@ -584,7 +584,7 @@ class People(fpb.BasePeople):
         '''
         Count how many total live women in each 5-year age bin 10-50, for tabulating ASFR
         '''
-        for key, (age_low, age_high) in fpd.age_bin_map.items():
+        for key, (age_low, age_high) in fpp.age_bin_map.items():
             this_age_bin = self.filter((self.age >= age_low) * (self.age < age_high))
             self.step_results['age_bin_totals'][key] += len(this_age_bin)
         return
@@ -665,7 +665,7 @@ class People(fpb.BasePeople):
             switching_postpartum = {}
         )
 
-        for key in fpd.age_bin_map.keys():
+        for key in fpp.age_bin_map.keys():
             self.step_results['birth_bins'][key] = 0
             self.step_results['age_bin_totals'][key] = 0
 
@@ -676,7 +676,7 @@ class People(fpb.BasePeople):
             return np.zeros((m, m), dtype=int)
 
         if self.pars['track_switching']:
-            for key in fpd.method_age_map.keys():
+            for key in fpp.method_age_map.keys():
                 self.step_results['switching_annual'][key]    = mm_zeros()
                 self.step_results['switching_postpartum'][key] = mm_zeros()
 
@@ -743,12 +743,12 @@ class Sim(fpb.BaseSim):
 
     def __init__(self, pars=None, location=None, label=None, mother_ids=False, **kwargs):
         if pars is None:
-            pars = fpd.pars(location)
+            pars = fpp.pars(location)
 
         # Check parameters
-        mismatches = [key for key in kwargs.keys() if key not in fpd.par_keys]
+        mismatches = [key for key in kwargs.keys() if key not in fpp.par_keys]
         if len(mismatches):
-            errormsg = f'Key(s) {mismatches} not found; available keys are {fpd.par_keys}'
+            errormsg = f'Key(s) {mismatches} not found; available keys are {fpp.par_keys}'
             raise sc.KeyNotFoundError(errormsg)
         super().__init__(pars, location=location, **kwargs) # Initialize and set the parameters as attributes
 
@@ -794,7 +794,7 @@ class Sim(fpb.BaseSim):
         self.results['birthday_fraction'] = []
         self.results['asfr'] = {}
 
-        for key in fpd.age_bin_map.keys():
+        for key in fpp.age_bin_map.keys():
             self.results['asfr'][key] = []
 
         if self['track_switching']:
@@ -1063,7 +1063,7 @@ class Sim(fpb.BaseSim):
             self.results['total_women_fecund'][i] = r.total_women_fecund*scale
             self.results['unintended_pregs'][i]   = r.unintended_pregs*scale
 
-            for agekey in fpd.age_bin_map.keys():
+            for agekey in fpp.age_bin_map.keys():
                 births_key = f'total_births_{agekey}'
                 women_key = f'total_women_{agekey}'
                 self.results[births_key][i] = r.birth_bins[agekey]*scale # Store results of total births per age bin for ASFR
@@ -1084,10 +1084,10 @@ class Sim(fpb.BaseSim):
                 self.results['switching_events_postpartum'][i] = scale*switch_events['postpartum']
 
             # Calculate metrics over the last year in the model and save whole years and stats to an array
-            if i % fpd.mpy == 0:
+            if i % fpp.mpy == 0:
                 self.results['tfr_years'].append(self.y)
-                start_index = (int(self.t)-1)*fpd.mpy
-                stop_index = int(self.t)*fpd.mpy
+                start_index = (int(self.t)-1)*fpp.mpy
+                stop_index = int(self.t)*fpp.mpy
                 unintended_pregs_over_year = scale*np.sum(self.results['unintended_pregs'][start_index:stop_index]) # Grabs sum of unintended pregnancies due to method failures over the last 12 months of calendar year
                 infant_deaths_over_year    = scale*np.sum(self.results['infant_deaths'][start_index:stop_index])
                 total_births_over_year     = scale*np.sum(self.results['total_births'][start_index:stop_index])
@@ -1113,7 +1113,7 @@ class Sim(fpb.BaseSim):
                     self.results['imr'].append(infant_mortality_rate)
 
                 tfr = 0
-                for key in fpd.age_bin_map.keys():
+                for key in fpp.age_bin_map.keys():
                     age_bin_births_year = np.sum(self.results['total_births_'+key][start_index:stop_index])
                     age_bin_total_women_year = self.results['total_women_'+key][stop_index]
                     age_bin_births_per_woman = sc.safedivide(age_bin_births_year, age_bin_total_women_year)
@@ -1270,7 +1270,7 @@ class Sim(fpb.BaseSim):
                     ax.fill_between(x, low, high, **fill_args)
 
             # Handle annotations
-            fpu.fixaxis(useSI=fpd.useSI, set_lim=new_fig) # If it's not a new fig, don't set the lim
+            fpu.fixaxis(useSI=fpp.useSI, set_lim=new_fig) # If it's not a new fig, don't set the lim
             if key == 'mcpr_by_year':
                 pl.ylabel('Percentage')
             elif key == 'mmr':
@@ -1364,7 +1364,7 @@ class Sim(fpb.BaseSim):
             pl.ylabel('Percentage')
             pl.xlabel('Year')
             pl.title(title, fontweight='bold')
-            fpu.fixaxis(useSI=fpd.useSI, set_lim=new_fig) # If it's not a new fig, don't set the lim
+            fpu.fixaxis(useSI=fpp.useSI, set_lim=new_fig) # If it's not a new fig, don't set the lim
 
         # Ensure the figure actually renders or saves
         sc.figlayout()