+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Version 0.12.0 (2022-05-21)
 --------------------------
 - Changed parameters from a dictionary to a class. This class has additional validation, the ability to import from/export to JSON, etc.
 - Added ability to add/remove contraceptive methods via ``pars.add_method()`` and ``pars.remove_method()``.
+- Renamed ``defaults.py`` to ``parameters.py``.
 - *GitHub info*: PR `503 <https://github.com/amath-idm/fp_analyses/pull/502>`__